+++ test/errors.ts
@@ -1,5 +1,8 @@
 import { assert } from 'chai';
 import { ApolloError } from '../src/errors/ApolloError';
+import HttpNetworkError from '../src/errors/HttpNetworkError';
+
+import { createFakeIResponse } from './mocks/mockFetch';
 
 describe('ApolloError', () => {
   it('should construct itself correctly', () => {
@@ -77,5 +80,33 @@ describe('ApolloError', () => {
       networkError,
     });
     assert(apolloError.stack, 'Does not contain a stack trace.');
+  });
+});
+
+describe('HttpNetworkError', () => {
+  it('should provide the given response back to the user', () => {
+    const response = createFakeIResponse('http://fake.url', 403, 'Unauthorized');
+
+    const err = new HttpNetworkError({ response });
+
+    assert.deepEqual(err.response, response);
+  });
+
+  it('should provide default values for the request and message', () => {
+    const response = createFakeIResponse('http://fake.url', 403, 'Unauthorized');
+    const err = new HttpNetworkError({ response });
+
+    assert.isOk(err.message);
+    assert.isObject(err.request);
+  });
+
+  it('should accept a request and message if provided', () => {
+    const response = createFakeIResponse('http://fake.url', 403, 'Unauthorized');
+    const request = { name: 'Sample Request' };
+    const message = 'a test message';
+    const err = new HttpNetworkError({ response, request, message });
+
+    assert.equal(err.message, message);
+    assert.deepEqual(err.request, request);
   });
 });