+++ test/mocks/mockFetch.ts
@@ -4,6 +4,7 @@ import 'whatwg-fetch';
 // structure to the MockedNetworkInterface.
 
 export interface MockedIResponse {
+  ok: boolean;
   json(): Promise<JSON>;
 }
 
@@ -14,8 +15,40 @@ export interface MockedFetchResponse {
   delay?: number;
 }
 
+export function createFakeIResponse(
+  url: string,
+  status: number,
+  statusText?: string,
+  body?: string,
+): IResponse {
+  return {
+    url,
+    status,
+    statusText,
+    ok: status === 200,
+    type: 'FakeResponse',
+    size: 0,
+    timeout: 0,
+    redirect: (_) => this,
+    error: () => this,
+    clone: () => this,
+    bodyUsed: !!body,
+    arrayBuffer: () => new Promise(resolve => resolve(new ArrayBuffer(0))),
+    blob: () => new Promise(resolve => resolve(new Blob())),
+    formData: () => new Promise(resolve => resolve(new FormData())),
+    json: () => new Promise(resolve => resolve({})),
+    text: () => new Promise(resolve => resolve('')),
+    headers: {
+      get: name => '',
+      getAll: name => [],
+      has: name => false,
+    },
+  };
+}
+
 export function createMockedIResponse(result: Object): MockedIResponse {
   return {
+    ok: true,
     json() {
       return Promise.resolve(result);
     },